updated/refactored unitTests, bash completion, assertError, and installer (unitTests is still a WIP) * added array* functions similar to map* and set* in bg_coreBashVars.sh * made mapSet support appending the the value with the -a and -d<delim> options * added *Size and *Clear functions to map* ans set* and array* * made setAdd support adding multiple entries on the cmdline * added -q|--quiet flag to returnValue to suppress writing to stdout if the varName does not exist * bgmktemp: supported -k|-keep flag on --release call as well as on the initial call. this facilitates easier testing * assertError: added catch_* variables for use in Catch: && { <errorPath...> } errorPath code. * fsTouch: added this as a simpler aaaTouch when you just need to know it exists. * added .bg-sp/onVinstall hook that installs the bootstrap library if needed. This had been hard coded in bg-debugCntr * bashCompletion: moved/renamed the bash completion library to use the new globalBashCompletion assetType * bashCompletion: started migrating to the more unique -hbOOBCompGen from -hb to make the mechanism more universal. * bashCompletion: bug fix: bgBCParse was not stopping on the = char * bashCompletion: created the bg_cmdChainLoader method of registering BC for cmds which eliminates the need for individual files for each command * unitTests: created whole new mechanism inspired by bg-scriptProjectDev but better in every way. *

+++ core/bg_coreFiles.sh
@@ -140,10 +140,11 @@ function fsMakeTemp()
 			local fileNameValue="${!fileNameVar}"
 		fi
 
+		local prevKeepFlag
 		local trapHandler="$(bgtrap -g -n "$fileNameValue"  EXIT 2>/dev/null)"
-		[[ "$trapHandler" =~ bgmktemp.*-k  ]] && keepFlag="1"
+		[[ "$trapHandler" =~ bgmktemp.*-k  ]] && prevKeepFlag="1"
 
-		if [ ! "$keepFlag" ]; then
+		if [ ! "$prevKeepFlag" ]; then
 			# remove the trap
 			bgtrap -r -n "$fileNameValue"  EXIT || assertError -c -v fileNameVar -v fileNameValue -V "$(trap -p EXIT)" "the trap previously set to remove the temp filecould not be removed."
 
@@ -154,7 +155,7 @@ function fsMakeTemp()
 			# TODO: decide if its safe to add an * to the end of this rm -rf or what we would have to do to safely rm
 			#       other temp files created by adding extensions to the base name. Some bg-lib code already does that
 			#       but maybe they should be refactored to create a temp directory to put multiple files
-			rm -rf "$fileNameValue"
+			[ ! "$keepFlag" ] && rm -rf "$fileNameValue"
 
 			if [ ! "$assertError_EndingScript" ] && [ "$mode" == "releaseInternal" ] && [ ! "$willNotReleaseFlag" ] && { [ "$BGMKTEMP_ERROR_UNRELEASED+exists" ] || bgtraceIsActive; }; then
 				(assertError -v fileNameValue -v trapHandler "a temp file created with bgmktemp was not released before the end of the script.")
